requirements.txt
streamlit==1.30.0
openai==1.12.0
python-dotenv==1.0.0
supabase==2.3.0
requests==2.31.0
pandas==2.1.4
matplotlib==3.8.2
plotly==5.18.0
pydantic==2.5.3
pytest==7.4.3
pytest-mock==3.12.0
python-dateutil==2.8.2
uuid==1.30
